Revert "fix perfil"
This reverts commit 0b2cc69706c0ed4c35e30b1aa5f6e5cd16a159ca.

+++ CookMate/Controllers/ProfileController.cs
@@ -26,7 +26,7 @@ namespace CookMate.Controllers {
         public IActionResult Delete(DeleteModel model) {
             int id = (int)HttpContext.Session.GetInt32("id");
             var user = _context.Utilizador.Find(id);
-            if (user.password == model.password) {
+            if (user != null && user.password == model.password) {
                 _context.Utilizador.Remove(user);
                 _context.SaveChanges();
                 return View("~/Views/Login/Login.cshtml");
